requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-cpp-python==0.3.7